--- Controls/searchNew.ts
@@ -1,4 +1,0 @@
-export {default as Controller} from './_searchNew/Controller';
-export {default as SearchInputContainer} from './_searchNew/InputContainer';
-export {default as SearchResolver} from './_searchNew/SearchResolver';
-export {ISearchResolverOptions} from './_searchNew/interface';
